Tidy up the README and clarify project "status"
The existing "project status" heading could be misunderstood as an indicator of release status. By instead using the phrase "production-ready" as an adjective in the short description, we are being more explicit that "production-ready" refers to the maturity of the codebase.

This change also tidies up the stylization of the CONTRIBUTING and LICENCE links.